src/main/java/algorithms/leetcode/MaximumNumberOfVowelsInASubstringOfGivenLength.java
package algorithms.leetcode;

/**
 * @link <a href="https://leetcode.com/problems/maximum-number-of-vowels-in-a-substring-of-given-length/">https://leetcode.com/problems/maximum-number-of-vowels-in-a-substring-of-given-length/</a>
 * @date 05.05.2023
 */
public class MaximumNumberOfVowelsInASubstringOfGivenLength {
    public int maxVowels(String s, int k) {
        int max = 0;
        int count = 0;
        for (int i = 0; i < k; i++) {
            if (isVowel(s.charAt(i))) {
                count++;
            }
        }
        max = count;
        for (int i = k; i < s.length(); i++) {
            if (isVowel(s.charAt(i))) {
                count++;
            }
            if (isVowel(s.charAt(i - k))) {
                count--;
            }
            max = Math.max(max, count);
        }
        return max;
    }

    private boolean isVowel(char ch) {
        return ch == 'a' || ch == 'e' || ch == 'i' || ch == 'o' || ch == 'u';
    }

}
